requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
pydantic==2.5.2
python-multipart==0.0.6
plaid-python==12.0.0
yfinance==0.2.28
ccxt==4.1.64
httpx==0.25.2
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.2
schedule==1.2.0
python-dateutil==2.8.2
fastapi-users==12.1.2
bcrypt==4.1.2
python-jose[cryptography]==3.3.0